Deleted the old cluster_service_2_x.py Renamed cluster_service_2_x_behaviors.py to cluster_service_2_x.py
