+++ container_service_extension/rde/backend/cluster_service_factory.py
@@ -7,8 +7,6 @@ import semantic_version
 import container_service_extension.common.utils.server_utils as server_utils
 from container_service_extension.rde.backend.cluster_service_1_x import ClusterService as ClusterService1X  # noqa: E501
 from container_service_extension.rde.backend.cluster_service_2_x import ClusterService as ClusterService2X  # noqa: E501
-from container_service_extension.rde.backend.cluster_service_2_x_behaviors import ClusterService as ClusterService2XBehaviors  # noqa: E501
-from container_service_extension.security.context.behavior_request_context import RequestContext  # noqa: E501
 
 
 class ClusterServiceFactory:
@@ -30,9 +28,5 @@ class ClusterServiceFactory:
         if rde_version.major == 1:
             return ClusterService1X(op_ctx=self.req_ctx)
         elif rde_version.major == 2:
-            # TODO This if else block needs to be removed once Behavior
-            #  integration is completed.
-            if isinstance(self.req_ctx, RequestContext):
-                return ClusterService2XBehaviors(self.req_ctx)
-            else:
-                return ClusterService2X(op_ctx=self.req_ctx)
+            return ClusterService2X(self.req_ctx)
+
